+++ examples/magnetized_sphere.py
@@ -22,8 +22,8 @@ loop_radius: float = 1.0  # (m)
 sphere_radius: float = 0.05  # (m)
 sphere_z: float = 0.2  # (m) height of the sphere above the xy plane
 mu_r: float = 1.5
-loop_mesh_size: float = 20e-3  # (m)
-sphere_mesh_sizes = [50e-3, 25e-3, 15e-3, 10e-3, 8e-3]
+loop_mesh_size: float = 50e-3  # (m)
+sphere_mesh_sizes = [30e-3, 20e-3, 15e-3, 10e-3, 8e-3]
 
 material = oersted.LinearMaterial(mu_r)
 
@@ -46,6 +46,16 @@ def make_loop_magnet(current: float, mesh_size: float):
     j_density[:, 0] = -jmag * np.sin(phi)
     j_density[:, 1] = jmag * np.cos(phi)
 
+    # oersted.plot_mesh(
+    #     loop_mesh,
+    #     filename="docs/figs/example_magnetized_sphere_j_density.svg",
+    #     centroids = loop_mesh.centroids,
+    #     vectors=j_density,
+    #     # scalars=np.linalg.norm(j_density, axis=1),
+    #     vector_scale = 5.0e-2,
+    #     transparency=True
+    # )
+
     return loop_mesh, j_density
 
 
@@ -64,7 +74,7 @@ def mesh_convergence_study(
         mesh.nodes[:, 2] += sphere_z
 
         # Plot both the sphere and loop together
-        # mesh.append(loop_mesh).plot()
+        # mesh.append(loop_mesh).plot("docs/figs/example_magnetized_sphere_meshes.svg")
 
         # Compute the external field acting on the sphere
         bext = oersted.b_field(loop_mesh, j_density, mesh.centroids, solver=solver)
@@ -97,10 +107,14 @@ def analytical_force() -> float:
     M: float = (1.0 / oersted.MU0) * b_centroid * chi
     volume: float = (4.0 * np.pi / 3.0) * sphere_radius**3
     m: float = M * volume
+    print(f"Magnetic field at sphere center: {b_centroid:.3f} T")
+    print(f"chi = {chi:.3f}")
+    print(f"Magnetization field, M: {M:.3f} A/m")
+    print(f"Magnetic moment, m = {m:.3f} A m^2")
 
     # Then, compute the field gradient
     dbdz = oersted.testing.dbzdz_loop_axis(current, loop_radius, sphere_z)
-
+    print(f"dBz/dz at sphere center: {dbdz:.3f} T/m")
     return m * dbdz
 
 
@@ -125,15 +139,45 @@ def plot_result(mesh_sizes, fx, fy, fz, fz_analytical):
     fig.savefig("docs/figs/example_magnetized_sphere_mesh_convergence.svg")
 
 
+def plot_fields_on_axis(loop_mesh: oersted.Mesh, j_density: NDArray[float64]):
+
+    n_pts: int = 100
+    z_max: float = 3.0
+    axis_pts = np.zeros((n_pts, 3))
+    axis_pts[:, 2] = np.linspace(0.0, z_max, n_pts)
+    bz_axis = oersted.b_field(loop_mesh, j_density, axis_pts, solver=solver)
+
+    n_pts_analytical: int = 20
+    axis_pts_analytical = np.linspace(0, z_max, n_pts_analytical)
+    bz_analytical = np.zeros((n_pts_analytical,))
+    for i in range(n_pts_analytical):
+        bz_analytical[i] = oersted.testing.bz_loop_axis(
+            current, loop_radius, axis_pts_analytical[i]
+        )
+
+    fig, ax = plt.subplots()
+    ax.plot(axis_pts[:, 2], bz_axis[:, 2], "k", label="oersted")
+    ax.plot(axis_pts_analytical, bz_analytical, "rs", label="analytical")
+    ax.set_xlabel("Distance Along Z-Axis (m)")
+    ax.set_ylabel("Magnetic Flux Density, B (T)")
+    ax.set_title("Magnetic Field Along the Loop Axis")
+    ax.legend()
+    fig.savefig("docs/figs/example_magnetized_sphere_field_on_axis.svg")
+
+
 def main():
     # Create the loop
     loop_mesh, j_density = make_loop_magnet(current, loop_mesh_size)
+
+    # Compute and plot the fields on axis
+    plot_fields_on_axis(loop_mesh, j_density)
 
     # Run the mesh convergence study
     fx, fy, fz = mesh_convergence_study(sphere_mesh_sizes, loop_mesh, j_density)
 
     # Compute analytical value
     fz_analytical = analytical_force()
+    print(f"Analytical force acting on the sphere: {fz_analytical:.3f} N")
 
     assert np.abs(fx[-1]) < 1.0
     assert np.abs(fy[-1]) < 1.0